refactor:code refactor based on the code review
