+++ src/components/ClientLayout.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import theme from '@/theme';
 
 import React from 'react';
