--- src/messages/en.json
@@ -1,14 +1,0 @@
-{
-  "common": {
-    "contactUs": "Contact Us",
-    "privacyPolicy": "Privacy Policy",
-    "termsForUsage": "Terms for Usage",
-    "wechat": "Wechat",
-    "copyright": "© 2025 Meetly Omni. All rights reserved."
-  },
-  "home": {
-    "title": "Meetly Omni Landing Page",
-    "underConstruction": "The page is under construction.",
-    "currentLocale": "Current locale: {locale}"
-  }
-}
