--- src/messages/zh.json
@@ -1,14 +1,0 @@
-{
-  "common": {
-    "contactUs": "联系我们",
-    "privacyPolicy": "隐私政策",
-    "termsForUsage": "使用条款",
-    "wechat": "微信",
-    "copyright": "© 2025 Meetly Omni. 保留所有权利。"
-  },
-  "home": {
-    "title": "Meetly Omni 落地页",
-    "underConstruction": "页面正在建设中。",
-    "currentLocale": "当前语言: {locale}"
-  }
-}
